Organize routes and controllers
